pyproject.toml
[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "floodmap"
version = "0.1.0"
description = "floodmap based on your location"
readme = "README.md"
requires-python = ">=3.12"
# Core dependencies for API serving (minimal runtime)
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.30.0",
    "numpy>=2.1.2",
    "Pillow[webp]>=10.3.0",
    "zstandard>=0.23.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-requests>=0.41b0",
    "requests>=2.32.4",
    "tqdm>=4.67.1",
    "brotli>=1.1.0",
    "cachetools>=5.3.0",
]

[project.optional-dependencies]
# Data processing dependencies (for offline data preparation)
processing = [
    "gdal==3.10.0",
    "rasterio>=1.4.1",
    "pyproj>=3.7.0",
    "matplotlib>=3.9.2",
    "scipy>=1.14.1",
    "tqdm>=4.66.5",
    "s3fs>=2024.3.1",
]

# Development dependencies (notebooks, analysis, etc.)
dev = [
    "diskcache>=5.6.3",
    "folium>=0.17.0",
    "geopy>=2.4.1",
    "googlemaps>=4.10.0",
    "ipykernel>=6.29.5",
    "python-fasthtml>=0.6.9",
    "redis>=5.0.4",
    "aiofiles>=23.0.0",
    "httptools>=0.6.0",
    "cryptography<46",
    "setuptools>=69.0.0",
    "wheel>=0.43.0",
    "lz4>=4.4.4",
    "prometheus-client>=0.20.0",
]

# Testing dependencies
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "playwright>=1.53.0",
    "pytest-playwright>=0.7.0",
]

[tool.setuptools]
py-modules = ["main"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "unit: fast unit tests (< 1s each)",
    "integration: integration tests that require external dependencies",
    "e2e: end-to-end tests using Playwright",
    "visual: visual regression tests",
    "performance: performance and load tests",
    "slow: slow tests that can be skipped in development",
    "asyncio: async tests",
    "aws: tests that require AWS credentials",
    "external: tests that require external resources like network access",
]
asyncio_mode = "auto"

[tool.ruff]
# Target Python 3.12
target-version = "py312"
line-length = 88

# Exclude auto-generated or vendor directories
exclude = [
    ".venv",
    ".git",
    "__pycache__",
    "*.egg-info",
]

[tool.ruff.lint]
# Enable specific ruff rules (replaces flake8 + additional modern checks)
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "SIM", # flake8-simplify
]

# Ignore specific rules that may be too strict for this codebase
ignore = [
    "E501",  # line-too-long (handled by formatter)
    "F841",  # unused-variable (allow for debugging/development)
    "E402",  # module-import-not-at-top (some scripts need this)
    "E722",  # bare-except (allow for now, can be fixed incrementally)
]

[tool.ruff.lint.isort]
known-first-party = ["src", "floodmap"]
force-single-line = false

[tool.ruff.format]
# Use double quotes for consistency
quote-style = "double"
# Prefer trailing commas in multi-line structures
skip-magic-trailing-comma = false
